Add files via upload
This Python script calculates the factorial of a given number using recursion. The factorial function takes an integer n as an argument and returns the factorial of n. The base cases are when n is 0 or 1, in which case the factorial is 1. For other values of n, the factorial is calculated recursively by multiplying n with the factorial of n - 1.

Use:
This script can be useful whenever you need to calculate the factorial of a number. It's a simple example of a recursive algorithm. You can upload this code to GitHub to showcase your coding skills, collaborate with others, or demonstrate your understanding of version control using Git and GitHub.